Add color-blind agents
This adds some "color-blind" agents to test the potential evolutionary
advantages of color awareness. This sets parameters so that the first
agent created is always a color-blind non-cooperator to see how
cooperation and color-awareness might evolve, and also increases the
cost of giving to examine circumstances in which cooperation with
non-cooperators might be more costly.