Tweak some parameters and add some extra interaction tests
I tweaked parameters in order to reproduce Axelrod and Hammond's result
that cooperation shows up much more strongly with 2 colors than with 1.
This doesn't exactly match their parameters, but it seems to work. An odd
thing was the strong initial rise in DD with these parameters, when usually
it's CC that makes the initial jump. Just to be on the safe side, I added
some extra interaction tests to make sure strategies were doing what they
said they were. They all passed. I am aware that these tests are a bit sloppy
and repetitious, but time is a factor here.